% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emissions_product}
\alias{emissions_product}
\title{Product level result of the emissions indicator}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 5 columns.
}
\usage{
emissions_product
}
\description{
Product level result of the emissions indicator
}
\examples{
emissions_product
}
\keyword{datasets}
