% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_transition_risk_profile_cols_company.R
\name{rename_transition_risk_profile_cols_company}
\alias{rename_transition_risk_profile_cols_company}
\title{Rename transition risk profile columns at company level}
\usage{
rename_transition_risk_profile_cols_company(data, include_co2 = FALSE)
}
\arguments{
\item{data}{Dataframe.}

\item{include_co2}{Logical. Include \verb{co2_*} columns ?}
}
\description{
Rename transition risk profile columns at company level
}
\keyword{internal}
