% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_transition_risk_profile_cols_product.R
\name{rename_transition_risk_profile_cols_product}
\alias{rename_transition_risk_profile_cols_product}
\title{Rename transition risk profile columns at product level}
\usage{
rename_transition_risk_profile_cols_product(data)
}
\arguments{
\item{data}{Dataframe.}
}
\description{
Rename transition risk profile columns at product level
}
\keyword{internal}
