% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_webtool_cols_at_company_level_wide.R
\name{rename_webtool_cols_at_company_level_wide}
\alias{rename_webtool_cols_at_company_level_wide}
\title{Rename webtool columns at company level for wider format}
\usage{
rename_webtool_cols_at_company_level_wide(data)
}
\arguments{
\item{data}{Dataframe.}
}
\description{
Rename webtool columns at company level for wider format
}
\keyword{internal}
