% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_webtool_cols_at_company_level_wide.R
\name{select_webtool_cols_at_company_level_wide}
\alias{select_webtool_cols_at_company_level_wide}
\title{Select webtool columns at company level for wider format}
\usage{
select_webtool_cols_at_company_level_wide(data)
}
\arguments{
\item{data}{Dataframe.}
}
\description{
Select webtool columns at company level for wider format
}
\keyword{internal}
