% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_webtool_cols_at_product_level.R
\name{select_webtool_cols_at_product_level}
\alias{select_webtool_cols_at_product_level}
\title{Select webtool columns at product level}
\usage{
select_webtool_cols_at_product_level(data)
}
\arguments{
\item{data}{Dataframe.}
}
\description{
Select webtool columns at product level
}
\keyword{internal}
