% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltPolish-package.R
\docType{package}
\name{tiltPolish-package}
\alias{tiltPolish}
\alias{tiltPolish-package}
\title{tiltPolish: Polish 'TILT' Ouput Data}
\description{
Polish 'TILT' output data to make it ready for presentation. This represents one of the external layers of software architecture of the 'TILT' ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/tiltPolish}
  \item \url{https://2degreesinvesting.github.io/tiltPolish/}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/tiltPolish/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

}
\keyword{internal}
