% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_transition_risk_profile_cols_company.R
\name{rename_transition_risk_profile_cols_company}
\alias{rename_transition_risk_profile_cols_company}
\title{Rename transition risk profile columns at company level}
\usage{
rename_transition_risk_profile_cols_company(data)
}
\arguments{
\item{data}{Dataframe.}
}
\description{
Rename transition risk profile columns at company level
}
\keyword{internal}
