% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_co2_range.R
\name{summarize_co2_range}
\alias{summarize_co2_range}
\title{Summarize the range of CO2}
\usage{
summarize_co2_range(data)
}
\arguments{
\item{data}{Data frame. The product level result of the emissions indicator.}
}
\value{
A data frame with the columns
benchmark, tilt_subsector, min, max and all rows
where the column benchmark equals tilt_subsector.
}
\description{
Summarize the range of CO2
}
\examples{
data <- emissions_product
summarize_co2_range(data)
}
