% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltPolish-package.R
\docType{package}
\name{tiltPolish-package}
\alias{tiltPolish}
\alias{tiltPolish-package}
\title{tiltPolish: Polish 'TILT' Ouput Data}
\description{
Polish 'TILT' output data to make it ready for presentation. This represents one of the external layers of software architecture of the 'TILT' ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/2DegreesInvesting/tiltPolish}
  \item \url{https://2degreesinvesting.github.io/tiltPolish/}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/tiltPolish/issues}
}

}
\author{
\strong{Maintainer}: Kalash Singhal \email{kalash@2degrees-investing.org}

Authors:
\itemize{
  \item Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})
  \item Anne Schoenauer \email{anne.schoenauer@2degrees-investing.org} (\href{https://orcid.org/0000-0002-4576-8799}{ORCID})
  \item Tilman Trompke \email{tilman@2degrees-investing.org} (\href{https://orcid.org/0009-0003-3351-5131}{ORCID})
  \item Linda Delacombaz \email{linda@2degrees-investing.org}
}

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
